refactor: reorganises files. renames feedItem file for consistent file naming
