+++ src/js/components/calendar-item.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import { View, Text, Image, TouchableOpacity } from 'react-native';
 import Icon from 'react-native-vector-icons/FontAwesome';
-import formatDate from '../../lib/format-date';
-import CardSection from '../common/CardSection';
-import Card from '../common/Card';
-import styles from '../../../styles';
+import formatDate from '../lib/format-date';
+import CardSection from './common/CardSection';
+import Card from './common/Card';
+import styles from '../../styles';
 
 /***
 * CalendarItem is used in calendar and album views. Plays equivalent role
@@ -63,7 +63,7 @@ const CalendarItem = ({ what, where, when,
 
             <Image
               style={styles.coverImage}
-              source={coverPhoto ? { uri: coverPhoto.photo_url } : require('../../../img/placeholder.png')}
+              source={coverPhoto ? { uri: coverPhoto.photo_url } : require('../../img/placeholder.png')}
             />
 
           </View>
