+++ src/js/components/feed-item.js
