+++ src/js/reducers/calendar.old.js
@@ -56,54 +56,19 @@ export default function calendar (state = initialState, action) {
   switch (action.type) {
 
     case actions.GET_CALENDAR_REQUEST:
-      return handleCalendarRequest(state, action);
+      return { ...state, isFetching: true };
 
     case actions.GET_CALENDAR_SUCCESS:
-      return handleCalendarSuccess(state, action);
+      return { ...state, isFetching: false, data: action.data };
 
     case actions.GET_CALENDAR_FAILURE:
-      return handleCalendarFailure(state, action);
+      return { ...state, isFetching: false, error: action.error };
 
     case actions.APPLY_FILTER:
     case actions.CLEAR_FILTER:
-      return handleFilter(state, action);
+      return { ...state, filter: action.filter, showHosting: action.showHosting };
 
     default:
       return state;
   }
-}
-
-function handleCalendarRequest (state, action) {
-
-  const newState = update(state, {
-    isFetching: { $set: action.isFetching }
-  });
-  return newState;
-}
-
-function handleCalendarSuccess (state, action) {
-
-  const newState = update(state, {
-    isFetching: { $set: action.isFetching },
-    data: { $set: action.data }
-  });
-  return newState;
-}
-
-function handleCalendarFailure (state, action) {
-
-  const newState = update(state, {
-    isFetching: { $set: action.isFetching },
-    error: { $set: action.error }
-  });
-  return newState;
-}
-
-function handleFilter (state, action) {
-
-  const newState = update(state, {
-    filter: { $set: action.filter },
-    showHosting: { $set: action.showHosting }
-  });
-  return newState;
 }
